+++ lib/benchpark/cmd/setup.py
@@ -9,7 +9,7 @@ import shutil
 import sys
 import yaml
 
-from benchpark.accounting import benchpark_experiments, benchpark_modifiers
+from benchpark.accounting import benchpark_experiments, benchpark_modifiers, benchpark_systems
 from benchpark.debug import debug_print
 from benchpark.paths import source_location
 from benchpark.runtime import RuntimeResources
