+++ lib/main.py
@@ -15,7 +15,7 @@ import yaml
 import benchpark.cmd.system
 import benchpark.cmd.experiment
 import benchpark.cmd.setup
-from benchpark.accounting import benchpark_experiments, benchpark_modifiers
+from benchpark.accounting import benchpark_experiments, benchpark_modifiers, benchpark_systems
 from benchpark.debug import debug_print
 from benchpark.paths import source_location
 from benchpark.runtime import RuntimeResources
@@ -84,17 +84,6 @@ def benchpark_benchmarks():
     for x in os.listdir(experiments_dir):
         benchmarks.append(f"{x}")
     return benchmarks
-
-
-def benchpark_systems():
-    source_dir = source_location()
-    systems = []
-    for x in os.listdir(source_dir / "configs"):
-        if not (
-            os.path.isfile(os.path.join(source_dir / "configs", x)) or x == "common"
-        ):
-            systems.append(x)
-    return systems
 
 
 def benchpark_get_tags():
